requirements.txt
fastapi
av
accelerate
config
cloudpickle
diffusers
einops
fvcore
gradio
gradio-client
imageio
matplotlib
numpy
omegaconf
onnxruntime
opencv-python
pandas
pillow
pycocotools
pyyaml
psutil
peft
regex==2024.5.15
safetensors
scikit-image
scipy
timm
tokenizers
torch
torchaudio
torchmetrics
torchvision
tqdm
transformers
triton